+++ CleantalkMod.php
@@ -13,21 +13,23 @@ if (!defined('SMF')) {
     die('Hacking attempt...');
 }
 
+// Fixes for old PHP versions
+require_once(dirname(__FILE__) . '/phpFix.php');
+
+// Base classes
 require_once(dirname(__FILE__) . '/Cleantalk.php');
 require_once(dirname(__FILE__) . '/CleantalkRequest.php');
 require_once(dirname(__FILE__) . '/CleantalkResponse.php');
 require_once(dirname(__FILE__) . '/CleantalkHelper.php');
 require_once(dirname(__FILE__) . '/CleantalkSFW.php');
 
-// define same CleanTalk options
-define('CT_AGENT_VERSION', 'smf-210');
+// Common CleanTalk options
+define('CT_AGENT_VERSION', 'smf-213');
 define('CT_SERVER_URL', 'http://moderate.cleantalk.org');
 define('CT_DEBUG', false);
 
 /**
  * CleanTalk SFW check
- * @param array $regOptions
- * @param array $theme_vars
  * @return void
  */
 function cleantalk_sfw_check()
@@ -35,11 +37,11 @@ function cleantalk_sfw_check()
 	global $modSettings, $user_info;
 		
 	if(!empty($modSettings['cleantalk_sfw']) && !$user_info['is_admin']){
-				
-		$sfw = new cleantalk\antispam\CleantalkSFW();
+		
+		$sfw = new CleantalkSFW();
 		$key = $modSettings['cleantalk_api_key'];
 		
-		$ips = $sfw->cleantalk_get_real_ip();
+		$ips = $sfw->get_ip();
 
 		$is_sfw_check=true;
 		foreach($ips as $curr_ip){
@@ -48,7 +50,7 @@ function cleantalk_sfw_check()
 				
 				$is_sfw_check=false;
 				if(!empty($_COOKIE['ct_sfw_passed'])){
-					$sfw->sfw_update_logs($sfw->passed_ip, 'passed');
+					$sfw->sfw_update_logs($curr_ip, 'passed');
 					setcookie ('ct_sfw_passed', '0', 1, "/");
 				}
 			}
@@ -82,10 +84,10 @@ function cleantalk_check_register(&$regOptions, $theme_vars){
     if ($regOptions['interface'] == 'admin')
         return;
 
-    $ct = new cleantalk\antispam\Cleantalk();
+    $ct = new Cleantalk();
     $ct->server_url = CT_SERVER_URL;
 
-    $ct_request = new cleantalk\antispam\CleantalkRequest();
+    $ct_request = new CleantalkRequest();
     $ct_request->auth_key = cleantalk_get_api_key();
 
     $ct_request->response_lang = 'en'; // SMF use any charset and language
@@ -169,126 +171,129 @@ function cleantalk_check_register(&$regOptions, $theme_vars){
 function cleantalk_check_message(&$msgOptions, $topicOptions, $posterOptions){
 	
     global $language, $user_info, $modSettings, $smcFunc, $db_connection;
-
+	
     if (SMF == 'SSI') {
 		return;
     }
-
-    if(!$modSettings['cleantalk_first_post_checking']){
-        return;
-    }elseif (isset($user_info['posts']) && $user_info['posts'] > 0){
-        return;
-    }
-
-    $ct = new cleantalk\antispam\Cleantalk();
-    $ct->server_url = CT_SERVER_URL;
-
-    $ct_request = new cleantalk\antispam\CleantalkRequest();
-    $ct_request->auth_key = cleantalk_get_api_key();
-
-    $ct_request->response_lang = 'en'; // SMF use any charset and language
-
-    $ct_request->agent = CT_AGENT_VERSION;
-    $ct_request->sender_email = isset($posterOptions['email']) ? $posterOptions['email'] : '';
-
-    $ip = isset($user_info['ip']) ? $user_info['ip'] : $_SERVER['REMOTE_ADDR'];
-    $ct_request->sender_ip = $ct->ct_session_ip($ip);
-
-    $ct_request->sender_nickname = isset($posterOptions['name']) ? $posterOptions['name'] : '';
-    $ct_request->message = $msgOptions['body'];
-
-    $ct_request->submit_time = cleantalk_get_form_submit_time();
-
-    $ct_request->js_on = cleantalk_is_valid_js() ? 1 : 0;
-
-    $ct_request->sender_info = json_encode(
-        array(
-            'REFFERRER'              => isset($_SERVER['HTTP_REFERER'])      ? $_SERVER['HTTP_REFERER']     : null,
-            'cms_lang'               => substr($language, 0, 2),                                            
-            'USER_AGENT'             => isset($_SERVER['HTTP_USER_AGENT'])   ? $_SERVER['HTTP_USER_AGENT']  : null,
-			'js_timezone'            => isset($_COOKIE['ct_timezone'])       ? $_COOKIE['ct_timezone']      : null,
-			'mouse_cursor_positions' => isset($_COOKIE['ct_pointer_data'])   ? $_COOKIE['ct_pointer_data']  : null,
-			'key_press_timestamp'    => !empty($_COOKIE['ct_fkp_timestamp']) ? $_COOKIE['ct_fkp_timestamp'] : null,
-			'page_set_timestamp'     => !empty($_COOKIE['ct_ps_timestamp'])  ? $_COOKIE['ct_ps_timestamp']  : null
-        )
-    );
-
-    if (isset($topicOptions['id'])) {
-		
-		if(!isset($db_connection) || $db_connection === false)
-				loadDatabase();
-			
-		if(isset($db_connection) && $db_connection != false){
-			
-				// disable query check for UNION operator
-				$oldQueryCheck = isset($modSettings['disableQueryCheck']) ? $modSettings['disableQueryCheck'] : false;
-				$modSettings['disableQueryCheck'] = true;
-
-				// find first and last 5 messages
-				$posts = $smcFunc['db_query'](
-					'',
-					'SELECT m.id_msg, m.body
-					   FROM {db_prefix}messages AS m
-					   JOIN {db_prefix}topics AS t ON t.id_first_msg=m.id_msg
-					   WHERE t.id_topic = {int:id_topic}
-					   UNION
-					   (SELECT m.id_msg, m.body
-					   FROM {db_prefix}messages AS m
-					   WHERE m.id_topic = {int:id_topic2} AND m.approved=1
-					   ORDER BY id_msg desc
-					   limit 5)
-					   ORDER BY id_msg',
-					array(
-							'id_topic' => $topicOptions['id'],
-							'id_topic2' => $topicOptions['id'],
-					)
-				);
-				$messages = array();
-				while ($post = $smcFunc['db_fetch_assoc']($posts)) {
-					$messages[] = $post['body'];
-				}
-				$smcFunc['db_free_result']($posts);
-				$modSettings['disableQueryCheck'] = $oldQueryCheck;
-
-				$ct_request->example = implode("\n", $messages);
+	
+	// Do not check admin
+	if(!$user_info['is_admin']){
+	
+		if(!$modSettings['cleantalk_first_post_checking']){
+			return;
+		}elseif (isset($user_info['posts']) && $user_info['posts'] > 0){
+			return;
 		}
-    }
 
-    if(defined('CT_DEBUG') && CT_DEBUG)
-        log_error('CleanTalk request: ' . var_export($ct_request, true), 'user');
+		$ct = new Cleantalk();
+		$ct->server_url = CT_SERVER_URL;
 
-    /**
-     * @var CleantalkResponse $ct_result CleanTalk API call result
-     */
-    $ct_result = $ct->isAllowMessage($ct_request);
-    $ct_answer_text = 'CleanTalk: ' . strip_tags($ct_result->comment);
+		$ct_request = new CleantalkRequest();
+		$ct_request->auth_key = cleantalk_get_api_key();
 
-    if($ct_result->errno != 0 && !cleantalk_is_valid_js()){
-        cleantalk_log('deny post (errno !=0, invalid js test)' . strip_tags($ct_result->comment));
-        fatal_error($ct_answer_text, false);
-        return;
-    }
+		$ct_request->response_lang = 'en'; // SMF use any charset and language
 
-    if ($ct_result->allow == 0){
-		$msgOptions['cleantalk_check_message_result'] = $ct_result->comment;
-		if ($modSettings['postmod_active']){
-            if ($ct_result->stop_queue == 1){
-                cleantalk_log('spam message "' . $ct_result->comment . '"');
-                fatal_error($ct_answer_text, false);
-            }else{
-                // If post moderation active then set message not approved
-                cleantalk_log('to postmoderation "' . $ct_result->comment . '"');
-                $msgOptions['approved'] = 0;
-            }
-        }else{
-            cleantalk_log('spam message "' . $ct_result->comment . '"');
-            fatal_error($ct_answer_text, false);
-        }
-    }else{
-        // all ok, only logging
-        cleantalk_log('allow message for "' . $posterOptions['name'] . '"');
-    }
-    
+		$ct_request->agent = CT_AGENT_VERSION;
+		$ct_request->sender_email = isset($posterOptions['email']) ? $posterOptions['email'] : '';
+
+		$ip = isset($user_info['ip']) ? $user_info['ip'] : $_SERVER['REMOTE_ADDR'];
+		$ct_request->sender_ip = $ct->ct_session_ip($ip);
+
+		$ct_request->sender_nickname = isset($posterOptions['name']) ? $posterOptions['name'] : '';
+		$ct_request->message = $msgOptions['body'];
+
+		$ct_request->submit_time = cleantalk_get_form_submit_time();
+
+		$ct_request->js_on = cleantalk_is_valid_js() ? 1 : 0;
+
+		$ct_request->sender_info = json_encode(
+			array(
+				'REFFERRER'              => isset($_SERVER['HTTP_REFERER'])      ? $_SERVER['HTTP_REFERER']     : null,
+				'cms_lang'               => substr($language, 0, 2),                                            
+				'USER_AGENT'             => isset($_SERVER['HTTP_USER_AGENT'])   ? $_SERVER['HTTP_USER_AGENT']  : null,
+				'js_timezone'            => isset($_COOKIE['ct_timezone'])       ? $_COOKIE['ct_timezone']      : null,
+				'mouse_cursor_positions' => isset($_COOKIE['ct_pointer_data'])   ? $_COOKIE['ct_pointer_data']  : null,
+				'key_press_timestamp'    => !empty($_COOKIE['ct_fkp_timestamp']) ? $_COOKIE['ct_fkp_timestamp'] : null,
+				'page_set_timestamp'     => !empty($_COOKIE['ct_ps_timestamp'])  ? $_COOKIE['ct_ps_timestamp']  : null
+			)
+		);
+
+		if (isset($topicOptions['id'])) {
+			
+			if(!isset($db_connection) || $db_connection === false)
+					loadDatabase();
+				
+			if(isset($db_connection) && $db_connection != false){
+				
+					// disable query check for UNION operator
+					$oldQueryCheck = isset($modSettings['disableQueryCheck']) ? $modSettings['disableQueryCheck'] : false;
+					$modSettings['disableQueryCheck'] = true;
+
+					// find first and last 5 messages
+					$posts = $smcFunc['db_query'](
+						'',
+						'SELECT m.id_msg, m.body
+						   FROM {db_prefix}messages AS m
+						   JOIN {db_prefix}topics AS t ON t.id_first_msg=m.id_msg
+						   WHERE t.id_topic = {int:id_topic}
+						   UNION
+						   (SELECT m.id_msg, m.body
+						   FROM {db_prefix}messages AS m
+						   WHERE m.id_topic = {int:id_topic2} AND m.approved=1
+						   ORDER BY id_msg desc
+						   limit 5)
+						   ORDER BY id_msg',
+						array(
+								'id_topic' => $topicOptions['id'],
+								'id_topic2' => $topicOptions['id'],
+						)
+					);
+					$messages = array();
+					while ($post = $smcFunc['db_fetch_assoc']($posts)) {
+						$messages[] = $post['body'];
+					}
+					$smcFunc['db_free_result']($posts);
+					$modSettings['disableQueryCheck'] = $oldQueryCheck;
+
+					$ct_request->example = implode("\n", $messages);
+			}
+		}
+
+		if(defined('CT_DEBUG') && CT_DEBUG)
+			log_error('CleanTalk request: ' . var_export($ct_request, true), 'user');
+
+		/**
+		 * @var CleantalkResponse $ct_result CleanTalk API call result
+		 */
+		$ct_result = $ct->isAllowMessage($ct_request);
+		$ct_answer_text = 'CleanTalk: ' . strip_tags($ct_result->comment);
+
+		if($ct_result->errno != 0 && !cleantalk_is_valid_js()){
+			cleantalk_log('deny post (errno !=0, invalid js test)' . strip_tags($ct_result->comment));
+			fatal_error($ct_answer_text, false);
+			return;
+		}
+
+		if ($ct_result->allow == 0){
+			$msgOptions['cleantalk_check_message_result'] = $ct_result->comment;
+			if ($modSettings['postmod_active']){
+				if ($ct_result->stop_queue == 1){
+					cleantalk_log('spam message "' . $ct_result->comment . '"');
+					fatal_error($ct_answer_text, false);
+				}else{
+					// If post moderation active then set message not approved
+					cleantalk_log('to postmoderation "' . $ct_result->comment . '"');
+					$msgOptions['approved'] = 0;
+				}
+			}else{
+				cleantalk_log('spam message "' . $ct_result->comment . '"');
+				fatal_error($ct_answer_text, false);
+			}
+		}else{
+			// all ok, only logging
+			cleantalk_log('allow message for "' . $posterOptions['name'] . '"');
+		}
+	}    
 }
 
 /**
@@ -386,14 +391,14 @@ function cleantalk_get_api_key(){
  */
 function cleantalk_general_mod_settings(&$config_vars){
 	
-    global $txt;
+    global $txt, $boardurl;
 	
     $config_vars[] = array('title', 'cleantalk_settings');
     $config_vars[] = array('text',  'cleantalk_api_key');
     $config_vars[] = array('check', 'cleantalk_first_post_checking');
-    $config_vars[] = array('check', 'cleantalk_logging');
-    $config_vars[] = array('check', 'cleantalk_tell_others', 'postinput' => $txt['cleantalk_tell_others_postinput']);
-    $config_vars[] = array('check', 'cleantalk_sfw', 'postinput' => $txt['cleantalk_sfw_postinput']);
+    $config_vars[] = array('check', 'cleantalk_logging', 'subtext' => sprintf($txt['cleantalk_logging_postinput'], $boardurl));
+    $config_vars[] = array('check', 'cleantalk_tell_others', 'subtext' => $txt['cleantalk_tell_others_postinput']);
+    $config_vars[] = array('check', 'cleantalk_sfw', 'subtext' => $txt['cleantalk_sfw_postinput']);
 	$config_vars[] = array('desc',  'cleantalk_api_key_description');
     $config_vars[] = array('desc',  'cleantalk_check_users');
 }
@@ -412,21 +417,28 @@ function cleantalk_print_js_input()
 			ctMouseEventTimerFlag = true, //Reading interval flag
 			ctMouseData = "[",
 			ctMouseDataCounter = 0;
-		
+
 		function ctSetCookie(c_name, value) {
 			document.cookie = c_name + "=" + encodeURIComponent(value) + "; path=/";
 		}
-		
+
 		ctSetCookie("ct_ps_timestamp", Math.floor(new Date().getTime()/1000));
 		ctSetCookie("ct_fkp_timestamp", "0");
 		ctSetCookie("ct_pointer_data", "0");
 		ctSetCookie("ct_timezone", "0");
-		
+
 		setTimeout(function(){
-			ctSetCookie("ct_timezone", d.getTimezoneOffset()/60*(-1));
 			ctSetCookie("ct_checkjs", "'.$value.'");
+			ctSetCookie("ct_timezone", d.getTimezoneOffset()/60*(-1));
 		},1000);
-		
+
+		//Writing first key press timestamp
+		var ctFunctionFirstKey = function output(event){
+			var KeyTimestamp = Math.floor(new Date().getTime()/1000);
+			ctSetCookie("ct_fkp_timestamp", KeyTimestamp);
+			ctKeyStopStopListening();
+		}
+
 		//Reading interval
 		var ctMouseReadInterval = setInterval(function(){
 				ctMouseEventTimerFlag = true;
@@ -437,8 +449,21 @@ function cleantalk_print_js_input()
 				var ctMouseDataToSend = ctMouseData.slice(0,-1).concat("]");
 				ctSetCookie("ct_pointer_data", ctMouseDataToSend);
 			}, 1200);
-		
-		//Stop observing function
+
+		//Logging mouse position each 300 ms
+		var ctFunctionMouseMove = function output(event){
+			if(ctMouseEventTimerFlag == true){
+				var mouseDate = new Date();
+				ctMouseData += "[" + Math.round(event.pageY) + "," + Math.round(event.pageX) + "," + Math.round(mouseDate.getTime() - ctTimeMs) + "],";
+				ctMouseDataCounter++;
+				ctMouseEventTimerFlag = false;
+				if(ctMouseDataCounter >= 100){
+					ctMouseStopData();
+				}
+			}
+		}
+
+		//Stop mouse observing function
 		function ctMouseStopData(){
 			if(typeof window.addEventListener == "function")
 				window.removeEventListener("mousemove", ctFunctionMouseMove);
@@ -447,19 +472,7 @@ function cleantalk_print_js_input()
 			clearInterval(ctMouseReadInterval);
 			clearInterval(ctMouseWriteDataInterval);				
 		}
-		
-		//Logging mouse position each 300 ms
-		var ctFunctionMouseMove = function output(event){
-			if(ctMouseEventTimerFlag == true){
-				var mouseDate = new Date();
-				ctMouseData += "[" + Math.round(event.pageY) + "," + Math.round(event.pageX) + "," + Math.round(mouseDate.getTime() - ctTimeMs) + "],";
-				ctMouseDataCounter++;
-				ctMouseEventTimerFlag = false;
-				if(ctMouseDataCounter >= 100)
-					ctMouseStopData();
-			}
-		}
-		
+
 		//Stop key listening function
 		function ctKeyStopStopListening(){
 			if(typeof window.addEventListener == "function"){
@@ -469,13 +482,6 @@ function cleantalk_print_js_input()
 				window.detachEvent("mousedown", ctFunctionFirstKey);
 				window.detachEvent("keydown", ctFunctionFirstKey);
 			}
-		}
-		
-		//Writing first key press timestamp
-		var ctFunctionFirstKey = function output(event){
-			var KeyTimestamp = Math.floor(new Date().getTime()/1000);
-			ctSetCookie("ct_fkp_timestamp", KeyTimestamp);
-			ctKeyStopStopListening();
 		}
 
 		if(typeof window.addEventListener == "function"){
@@ -531,6 +537,7 @@ function cleantalk_load()
 		return;
     }
 	
+	// Output JS for users
     if (
         isset($context['template_layers']) &&
         is_array($context['template_layers']) &&
@@ -540,138 +547,119 @@ function cleantalk_load()
     ) {
         $context ['html_headers'] .= cleantalk_print_js_input();
     }
-
-	// Getting key automatically
-	if(isset($_GET['ctgetautokey'])){
+	
+	// Only for admin
+	if(!empty($user_info['is_admin'])){
 		
-		$result = cleantalk\antispam\CleantalkHelper::getAutoKey($user_info['email'], $_SERVER['SERVER_NAME'], 'smf', 'antispam');
+		// Getting key automatically
+		if(!empty($_GET['ctgetautokey'])){
+			
+			$result = CleantalkHelper::getApiKey($user_info['email'], $_SERVER['SERVER_NAME'], 'smf');
+			
+			if (empty($result['error'])){
+							
+				// Doing noticePaidTill(), sfw update and sfw send logs via cron
+				$doing_cron = true;
+				
+				$settings_array = array(
+					'cleantalk_api_key_is_ok' => '1',
+					'cleantalk_api_key'       => $result['auth_key'],
+					'cleantalk_user_token'    => empty($result['user_token']) ? '' : $result['user_token']
+				);
+				
+				updateSettings($settings_array, false);
+				
+			}
+		}
 		
-		if (empty($result['error'])){
+		// Settings saved
+		if(isset($_POST['cleantalk_api_key'])){
 			
-			updateSettings(array('cleantalk_api_key_is_ok' => '1'), false);
-			updateSettings(array('cleantalk_api_key' => $result['auth_key']), false);
+			checkSession('request');
 			
-			// Doing noticePaidTill(), sfw update and sfw send logs via cron
-			$modSettings['cleantalk_api_key'] = $result['auth_key'];
-			$modSettings['cleantalk_api_key_is_ok'] = 1;
-			$modSettings['cleantalk_last_account_check'] = time()-10;
-			$modSettings['cleantalk_sfw_last_update'] = time()-10;
-			$modSettings['cleantalk_sfw_last_update'] = time()-10;
-
-			// User token is empty, request it via noticePaidTill()
-			if (empty($result['user_token'])){
-				$result = cleantalk\antispam\CleantalkHelper::noticePaidTill($result['auth_key']);
-				if (empty($result['error']))
-					updateSettings(array('cleantalk_user_token' => $result['user_token']), false);
+			// Send feedback about version
+			$ct = new Cleantalk();
+			$ct->server_url = CT_SERVER_URL;
+			$ct_request = new CleantalkRequest();
+			$ct_request->auth_key = cleantalk_get_api_key();
+			$ct_request->feedback = '0:'.CT_AGENT_VERSION;
+			$ct_result = $ct->sendFeedback($ct_request);
+			// unset($ct, $ct_request);
+			
+			// Check if key is valid
+			if($_POST['cleantalk_api_key'] != $modSettings['cleantalk_api_key']){
+				
+				$key_to_validate = strval($_POST['cleantalk_api_key']);
+				$result = CleantalkHelper::noticeValidateKey($key_to_validate);
+				
+				if(empty($result['error'])){
+					
+					if(!empty($result['valid'])){
+						
+						updateSettings(
+							array(
+								'cleantalk_api_key_is_ok' => '1',
+								'cleantalk_api_key' => $key_to_validate								
+							),
+							false
+						);
+						
+					}
+					else
+						updateSettings(array('cleantalk_api_key_is_ok' => '0', false));
+				}
 			}
 			
-		}
-	}
-	
-	// Check if key is valid
-	if(isset($_POST['cleantalk_api_key']) && $_POST['cleantalk_api_key'] != $modSettings['cleantalk_api_key']){
-		
-		$key_to_validate = strval($_POST['cleantalk_api_key']);
-		
-		$result = cleantalk\antispam\CleantalkHelper::noticeValidateKey($key_to_validate);
-		
-		if(empty($result['error'])){
+			// If key is valid doing noticePaidTill(), sfw update and sfw send logs via cron
+			if(!empty($modSettings['cleantalk_api_key_is_ok']))
+				$doing_cron = true;
 			
-			if($result && isset($result['valid']) && intval($result['valid']) == 1){
-				updateSettings(array('cleantalk_api_key_is_ok' => '1', false));
-				
-				// If key is valid doing noticePaidTill(), sfw update and sfw send logs via cron
-				$modSettings['cleantalk_api_key'] = $key_to_validate;
-				$modSettings['cleantalk_api_key_is_ok'] = 1;
-				$modSettings['cleantalk_last_account_check'] = time()-10;
-				$modSettings['cleantalk_sfw_last_update'] = time()-10;
-				$modSettings['cleantalk_sfw_last_update'] = time()-10;
+		}
+		
+		// Deleting selected users
+		if(isset($_POST['ct_del_user']))
+		{			
+			checkSession('request');
+			
+			if (!isset($db_connection) || $db_connection === false)
+				loadDatabase();
+			
+			if (isset($db_connection) && $db_connection != false)
+			{
+				foreach($_POST['ct_del_user'] as $key=>$value)
+				{
+					$result = $smcFunc['db_query']('', 'delete from {db_prefix}members where id_member='.intval($key),Array('db_error_skip' => true));
+					$result = $smcFunc['db_query']('', 'delete from {db_prefix}topics where id_member_started='.intval($key),Array('db_error_skip' => true));
+					$result = $smcFunc['db_query']('', 'delete from {db_prefix}messages where id_member='.intval($key),Array('db_error_skip' => true));
+				}
 			}
-			else
-				updateSettings(array('cleantalk_api_key_is_ok' => '0', false));
-			
 		}
 		
-		
+		// Deleting all users
+		if(isset($_POST['ct_delete_all']))
+		{
+			checkSession('request');
+			
+			if (!isset($db_connection) || $db_connection === false)
+				loadDatabase();
+			
+			if (isset($db_connection) && $db_connection != false)
+			{
+				$result = $smcFunc['db_query']('', 'select * from {db_prefix}members where ct_marked=1',Array());
+				while($row = $smcFunc['db_fetch_assoc'] ($result))
+				{
+					$tmp = $smcFunc['db_query']('', 'delete from {db_prefix}topics where id_member_started='.$row['id_member'],Array('db_error_skip' => true));
+					$tmp = $smcFunc['db_query']('', 'delete from {db_prefix}messages where id_member='.$row['id_member'],Array('db_error_skip' => true));
+				}
+				$result = $smcFunc['db_query']('', 'delete from {db_prefix}members where ct_marked=1',Array('db_error_skip' => true));
+			}
+		}
 	}
-	
-    if($user_info['is_admin'] && isset($_POST['ct_del_user'])){	
-		
-		checkSession('request');
-		
-		if (!isset($db_connection) || $db_connection === false)
-		    loadDatabase();
-		
-		if (isset($db_connection) && $db_connection != false){
-		    foreach($_POST['ct_del_user'] as $key=>$value){
-				
-				$result = $smcFunc['db_query']('', 'delete from {db_prefix}members where id_member='.intval($key),Array('db_error_skip' => true));
-				$result = $smcFunc['db_query']('', 'delete from {db_prefix}topics where id_member_started='.intval($key),Array('db_error_skip' => true));
-				$result = $smcFunc['db_query']('', 'delete from {db_prefix}messages where id_member='.intval($key),Array('db_error_skip' => true));
-				
-		    }
-		}
-    }
-
-    if($user_info['is_admin'] && isset($_POST['ct_delete_all'])){
-		
-		checkSession('request');
-		
-		if (!isset($db_connection) || $db_connection === false)
-		    loadDatabase();
-		
-		if (isset($db_connection) && $db_connection != false){
-			
-		    $result = $smcFunc['db_query']('', 'select * from {db_prefix}members where ct_marked=1',Array());
-			
-		    while($row = $smcFunc['db_fetch_assoc'] ($result)){
-				$tmp = $smcFunc['db_query']('', 'delete from {db_prefix}topics where id_member_started='.$row['id_member'],Array('db_error_skip' => true));
-				$tmp = $smcFunc['db_query']('', 'delete from {db_prefix}messages where id_member='.$row['id_member'],Array('db_error_skip' => true));
-		    }
-			
-		    $result = $smcFunc['db_query']('', 'delete from {db_prefix}members where ct_marked=1',Array('db_error_skip' => true));
-		}
-    }
-
-	// add "tell others" templates
-    if (isset($context['template_layers'])
-        && $context['template_layers'] === array('html', 'body')
-        && array_key_exists('cleantalk_tell_others', $modSettings)
-        && $modSettings['cleantalk_tell_others']
-    ){
-        $context['template_layers'][] = 'cleantalk';
-	}
-	
-    if($user_info['is_admin'] && isset($_POST['cleantalk_api_key'])){
-
-		checkSession('request');
-
-    	$ct = new cleantalk\antispam\Cleantalk();
-        $ct->server_url = CT_SERVER_URL;
-     
-        $ct_request = new cleantalk\antispam\CleantalkRequest();
-        $ct_request->auth_key = cleantalk_get_api_key();
-		$ct_request->feedback = '0:'.CT_AGENT_VERSION;
-		
-		$ct_result = $ct->sendFeedback($ct_request);		
-    }
-	
-	/* Update SFW and send logs if settings seved */
-    if($user_info['is_admin'] && isset($_POST['cleantalk_sfw']) && (int)$_POST['cleantalk_sfw'] == 1){
-
-		$sfw = new cleantalk\antispam\CleantalkSFW;
-		$sfw->sfw_update($modSettings['cleantalk_api_key']);
-		$sfw->send_logs($modSettings['cleantalk_api_key']);
-		unset($sfw);
-		updateSettings(array('cleantalk_sfw_last_update' => time()+86400), false);
-		updateSettings(array('cleantalk_sfw_last_logs_sent' => time()+3600), false);
-		
-    }
 	
 	/* Cron for update SFW */
-	if(!empty($modSettings['cleantalk_api_key_is_ok']) && !empty($modSettings['cleantalk_sfw']) && isset($modSettings['cleantalk_sfw_last_update']) && $modSettings['cleantalk_sfw_last_update'] < time()){
+	if(!empty($modSettings['cleantalk_api_key_is_ok']) && !empty($modSettings['cleantalk_sfw']) && isset($modSettings['cleantalk_sfw_last_update']) && $modSettings['cleantalk_sfw_last_update'] < time() || !empty($doing_cron)){
 
-		$sfw = new cleantalk\antispam\CleantalkSFW;
+		$sfw = new CleantalkSFW;
 		$sfw->sfw_update($modSettings['cleantalk_api_key']);
 		unset($sfw);
 		updateSettings(array('cleantalk_sfw_last_update' => time()+86400), false);
@@ -679,9 +667,9 @@ function cleantalk_load()
 	}
 	
 	/* Cron for send SFW logs */
-	if(!empty($modSettings['cleantalk_api_key_is_ok']) && !empty($modSettings['cleantalk_sfw']) && isset($modSettings['cleantalk_sfw_last_logs_sent']) && $modSettings['cleantalk_sfw_last_logs_sent'] < time()){
+	if(!empty($modSettings['cleantalk_api_key_is_ok']) && !empty($modSettings['cleantalk_sfw']) && isset($modSettings['cleantalk_sfw_last_logs_sent']) && $modSettings['cleantalk_sfw_last_logs_sent'] < time() || !empty($doing_cron)){
 		
-		$sfw = new cleantalk\antispam\CleantalkSFW;
+		$sfw = new CleantalkSFW;
 		$sfw->send_logs($modSettings['cleantalk_api_key']);
 		unset($sfw);
 		updateSettings(array('cleantalk_sfw_last_logs_sent' => time()+3600), false);
@@ -689,30 +677,39 @@ function cleantalk_load()
 	}
 	
 	/* Cron for account status */
-	if(!empty($modSettings['cleantalk_api_key_is_ok']) && isset($modSettings['cleantalk_last_account_check']) && $modSettings['cleantalk_last_account_check'] < time()){
+	if(!empty($modSettings['cleantalk_api_key_is_ok']) && isset($modSettings['cleantalk_last_account_check']) && $modSettings['cleantalk_last_account_check'] < time() || !empty($doing_cron)){
 		
-		$result = cleantalk\antispam\CleantalkHelper::noticePaidTill($modSettings['cleantalk_api_key']);
+		$result = CleantalkHelper::noticePaidTill($modSettings['cleantalk_api_key']);
 		
 		if(empty($result['error'])){
 			$settings_array = array(
-				'cleantalk_show_notice' => $result['show_notice'],
-				'cleantalk_renew'       => $result['renew'],      
-				'cleantalk_trial'       => $result['trial'],      
-				'cleantalk_user_token'  => $result['user_token'], 
-				'cleantalk_spam_count'  => $result['spam_count'], 
-				'cleantalk_moderate_ip' => $result['moderate_ip'],
-				'cleantalk_show_review' => $result['show_review'],
-				'cleantalk_ip_license'  => $result['ip_license'],
-				'cleantalk_last_account_check'   => time()+86400
+				'cleantalk_show_notice'        => (int)    $result['show_notice'],
+				'cleantalk_renew'              => (int)    $result['renew'],      
+				'cleantalk_trial'              => (int)    $result['trial'],      
+				'cleantalk_user_token'         => (string) $result['user_token'], 
+				'cleantalk_spam_count'         => (int)    $result['spam_count'], 
+				'cleantalk_moderate_ip'        => (int)    $result['moderate_ip'],
+				'cleantalk_show_review'        => (int)    $result['show_review'],
+				'cleantalk_ip_license'         => (int)    $result['ip_license'],
+				'cleantalk_last_account_check' => time()+86400
 			);
 		}else{
 			$settings_array = array(
-				'cleantalk_last_account_check'   => time()+3600
+				'cleantalk_last_account_check' => time()+3600
 			);
 		}
 		
 		updateSettings($settings_array, false);
 		
+	}
+	
+	// Add "tell others" templates
+    if (isset($context['template_layers'])
+        && $context['template_layers'] === array('html', 'body')
+        && array_key_exists('cleantalk_tell_others', $modSettings)
+        && $modSettings['cleantalk_tell_others']
+    ){
+        $context['template_layers'][] = 'cleantalk';
 	}
 }
 
@@ -761,11 +758,11 @@ function cleantalk_is_valid_js()
  */
 function template_cleantalk_above()
 {
-	global $user_info, $modSettings, $txt;
+	global $user_info, $modSettings, $txt, $boardurl;
 	
 	if($user_info['is_admin'] && isset($_GET['action']) && $_GET['action'] == 'admin'){
 		
-		$source_dir = (empty($_SERVER['HTTPS']) ? 'http://' : 'https://') . $_SERVER['HTTP_HOST'] . '/Sources/Cleantalk/';
+		$source_dir = $boardurl . '/Sources/Cleantalk/';
 		
 		echo "<div class='notice_wrapper'>";
 		
@@ -800,7 +797,8 @@ function template_cleantalk_above()
 			// }
 			
 			// Bad key banner
-			if(empty($modSettings['cleantalk_api_key_is_ok']) && isset($_GET['area']) && $_GET['area'] != 'modsettings'){
+			if(empty($modSettings['cleantalk_api_key_is_ok']) && (empty($_GET['area'])  || (isset($_GET['area'])&& $_GET['area'] != 'modsettings'))){
+				
 				echo "<div style='margin-bottom: 1.2em; padding: 5px;'>"
 					."<h1>"
 						."<img style='height: 20px; margin: 0 5px 0 0; position: relative; top: 5px;' src='{$source_dir}attention.png' />"
@@ -838,7 +836,7 @@ function cleantalk_buffer($buffer)
 	if (SMF == 'SSI')
 	    return $buffer;
 
-	if($user_info['is_admin'] && isset($_GET['action'], $_GET['area']) && $_GET['action'] == 'admin' && $_GET['area'] == 'modsettings'){
+	if($user_info['is_admin'] && isset($_GET['action'], $_GET['area']) && $_GET['action'] == 'admin' && $_GET['area'] == 'modsettings'){// && false){
 		
 		if(strpos($forum_version, 'SMF 2.0')===false){
 			
@@ -859,182 +857,188 @@ function cleantalk_buffer($buffer)
 		
 		if (isset($db_connection) && $db_connection != false){
 			
-			db_extend('packages');
-			
-		    $cols = $smcFunc['db_list_columns'] ('{db_prefix}members', 0);
-			
-		    if(in_array('ct_marked', $cols)){
+			if(isset($_GET['ctcheckspam'])){
 				
-				if(isset($_GET['ctcheckspam'])){
-					$sql = 'UPDATE {db_prefix}members set ct_marked=0';
-					$result = $smcFunc['db_query']('', $sql, Array());
-					$sql = 'SELECT * FROM {db_prefix}members where passwd<>""';
-					$result = $smcFunc['db_query']('', $sql, Array());
-					$users=Array();
-					$users[0]=Array();
-					$data=Array();
-					$data[0]=Array();
-					$cnt=0;
-					while($row = $smcFunc['db_fetch_assoc'] ($result)){
-							
-						$users[$cnt][] = array(
-							'name' => $row['member_name'],
-							'id' => $row['id_member'],
-							'email' => $row['email_address'],
-							'ip' => $row['member_ip'],
-							'joined' => $row['date_registered'],
-							'visit' => $row['last_login'],
-						);
-						
-						$data[$cnt][]=$row['email_address'];
-						$data[$cnt][]=$row['member_ip'];
-						
-						if(sizeof($users[$cnt])>450){
-							$cnt++;
-							$users[$cnt]=Array();
-							$data[$cnt]=Array();
-						}
-					}
-				
-					$error="";
-				
-					for($i=0;$i<sizeof($users);$i++){
-						$send=implode(',',$data[$i]);
-						$req="data=$send";
-						$opts = array(
-							'http'=>array(
-								'method'=>"POST",
-								'content'=>$req,
-							)
-						);
-						
-						$context = stream_context_create($opts);
-						$result = @file_get_contents("https://api.cleantalk.org/?method_name=spam_check_cms&auth_key=".cleantalk_get_api_key(), 0, $context);
-						$result=json_decode($result);
-						if(isset($result->error_message)){
-							$error=$result->error_message;
-						}else{
-							
-							if(isset($result->data)){
+				if($modSettings['cleantalk_api_key_is_ok'] == '1' && $modSettings['cleantalk_api_key'] != ''){
+					
+					db_extend('packages');
+					
+					$cols = $smcFunc['db_list_columns'] ('{db_prefix}members', 0);
+					
+					if(in_array('ct_marked', $cols)){
+					
+						$sql = 'UPDATE {db_prefix}members set ct_marked=0';
+						$result = $smcFunc['db_query']('', $sql, Array());
+						$sql = 'SELECT * FROM {db_prefix}members where passwd<>""';
+						$result = $smcFunc['db_query']('', $sql, Array());
+						$users=Array();
+						$users[0]=Array();
+						$data=Array();
+						$data[0]=Array();
+						$cnt=0;
+						while($row = $smcFunc['db_fetch_assoc'] ($result)){
 								
-								foreach($result->data as $key=>$value){
+							$users[$cnt][] = array(
+								'name' => $row['member_name'],
+								'id' => $row['id_member'],
+								'email' => $row['email_address'],
+								'ip' => $row['member_ip'],
+								'joined' => $row['date_registered'],
+								'visit' => $row['last_login'],
+							);
+							
+							$data[$cnt][]=$row['email_address'];
+							$data[$cnt][]=$row['member_ip'];
+							
+							if(sizeof($users[$cnt])>450){
+								$cnt++;
+								$users[$cnt]=Array();
+								$data[$cnt]=Array();
+							}
+						}
+					
+						$error="";
+					
+						for($i=0;$i<sizeof($users);$i++){
+							$send=implode(',',$data[$i]);
+							$req="data=$send";
+							$opts = array(
+								'http'=>array(
+									'method'=>"POST",
+									'content'=>$req,
+								)
+							);
+							
+							$context = stream_context_create($opts);
+							$result = @file_get_contents("https://api.cleantalk.org/?method_name=spam_check_cms&auth_key=".cleantalk_get_api_key(), 0, $context);
+							$result=json_decode($result);
+							if(isset($result->error_message)){
+								$error=$result->error_message;
+							}else{
+								
+								if(isset($result->data)){
 									
-									if($key === filter_var($key, FILTER_VALIDATE_IP)){
+									foreach($result->data as $key=>$value){
 										
-										if($value->appears==1){
+										if($key === filter_var($key, FILTER_VALIDATE_IP)){
 											
-											$sql = 'UPDATE {db_prefix}members set ct_marked=1 where member_ip="'.$key.'"';
-											$result = $smcFunc['db_query']('', $sql, Array('db_error_skip' => true));
-										}
-										
-									}else{
-										
-										if($value->appears==1){
-											$sql = 'UPDATE {db_prefix}members set ct_marked=1 where email_address="'.$key.'"';
-											$result = $smcFunc['db_query']('', $sql, Array('db_error_skip' => true));
+											if($value->appears==1){
+												
+												$sql = 'UPDATE {db_prefix}members set ct_marked=1 where member_ip="'.$key.'"';
+												$result = $smcFunc['db_query']('', $sql, Array('db_error_skip' => true));
+											}
+											
+										}else{
+											
+											if($value->appears==1){
+												$sql = 'UPDATE {db_prefix}members set ct_marked=1 where email_address="'.$key.'"';
+												$result = $smcFunc['db_query']('', $sql, Array('db_error_skip' => true));
+											}
 										}
 									}
 								}
 							}
 						}
-					}
-				
-					if($error!='')
-						$html.='<center>'
-									.'<div style="border:2px solid red;color:red;font-size:16px;width:300px;padding:5px;">'
-										.'<b>'.$error.'</b>'
-									.'</div>'
-									.'<br>'
-								.'</center>';
-								
-				}
-				
-				$sql = "SELECT * FROM {db_prefix}members WHERE ct_marked=1";
-				$result = $smcFunc['db_query']('', $sql, Array());
-				
-				if($smcFunc['db_num_rows'] ($result) == 0 && isset($_GET['ctcheckspam'])){
 					
-					$html.='<center><div><b>'.$txt['cleantalk_check_users_nofound'].'</b></div><br><br></center>';
-					
-				}else if($smcFunc['db_num_rows'] ($result) > 0){
-					
-					if(isset($_GET['ctcheckspam'])){
-						$html.='<center><h3>'.$txt['cleantalk_check_users_done'].'</h3><br /><br /></center>';
+						if($error!='')
+							$html.='<center>'
+										.'<div style="border:2px solid red;color:red;font-size:16px;width:300px;padding:5px;">'
+											.'<b>'.$error.'</b>'
+										.'</div>'
+										.'<br>'
+									.'</center>';
+									
 					}
 					
-					// Pagination			
-					$on_page = 20;
-					$pages = ceil(intval($smcFunc['db_num_rows'] ($result))/$on_page);
-					$page = !empty($_GET['spam_page']) ? intval($_GET['spam_page']) : 1;
-					$offset = ($page-1)*$on_page;
-					
-					$sql = "SELECT * FROM {db_prefix}members WHERE ct_marked=1 LIMIT $offset, $on_page";
+					$sql = "SELECT * FROM {db_prefix}members WHERE ct_marked=1";
 					$result = $smcFunc['db_query']('', $sql, Array());
 					
-					if($pages && $pages != 1){
-						$html.= "<div style='margin: 10px;'>"
-								."<b>".$txt['cleantalk_check_users_pages'].":</b>"
-								."<ul style='display: inline-block; margin: 0; padding: 0;'>";
-									for($i = 1; $i <= $pages; $i++){
-										$html.= "<li style='display: inline-block;	margin-left: 10px;'>"
-												."<a href='".preg_replace('/(&ctcheckspam=.*|&spam_page=.*)/', '', $_SERVER['REQUEST_URI'])."&spam_page=$i'>"
-													.($i == $page ? "<span style='font-size: 1.1em; font-weight: 600;'>$i</span>" : $i)
-												."</a>"
-											."</li>";
-									}
-							$html.= "</ul>";
-						$html.= "</div>";
-					}
-					
-					$html.='<center><table style="border-color:#666666;" border=1 cellspacing=0 cellpadding=3>'
-						.'<thead>'
-						.'<tr>'
-							.'<th>'.$txt['cleantalk_check_users_tbl_select'].'</th>'
-							.'<th>'.$txt['cleantalk_check_users_tbl_username'].'</th>'
-							.'<th>'.$txt['cleantalk_check_users_tbl_joined'].'</th>'
-							.'<th>E-mail</th>'
-							.'<th>IP</th>'
-							.'<th>'.$txt['cleantalk_check_users_tbl_lastvisit'].'</th>'
-							.'<th>'.$txt['cleantalk_check_users_tbl_posts'].'</th>'
-						.'</tr>'
-						.'</thead>'
-						.'<tbody>';
+					if($smcFunc['db_num_rows'] ($result) == 0 && isset($_GET['ctcheckspam'])){
 						
-					$found=false;
-					
-					while($row = $smcFunc['db_fetch_assoc'] ($result)){
+						$html.='<center><div><b>'.$txt['cleantalk_check_users_nofound'].'</b></div><br><br></center>';
+						
+					}else if($smcFunc['db_num_rows'] ($result) > 0){
+						
+						if(isset($_GET['ctcheckspam'])){
+							$html.='<center><h3>'.$txt['cleantalk_check_users_done'].'</h3><br /><br /></center>';
+						}
+						
+						// Pagination			
+						$on_page = 20;
+						$pages = ceil(intval($smcFunc['db_num_rows'] ($result))/$on_page);
+						$page = !empty($_GET['spam_page']) ? intval($_GET['spam_page']) : 1;
+						$offset = ($page-1)*$on_page;
+						
+						$sql = "SELECT * FROM {db_prefix}members WHERE ct_marked=1 LIMIT $offset, $on_page";
+						$result = $smcFunc['db_query']('', $sql, Array());
+						
+						if($pages && $pages != 1){
+							$html.= "<div style='margin: 10px;'>"
+									."<b>".$txt['cleantalk_check_users_pages'].":</b>"
+									."<ul style='display: inline-block; margin: 0; padding: 0;'>";
+										for($i = 1; $i <= $pages; $i++){
+											$html.= "<li style='display: inline-block;	margin-left: 10px;'>"
+													."<a href='".preg_replace('/(&ctcheckspam=.*|&spam_page=.*)/', '', $_SERVER['REQUEST_URI'])."&spam_page=$i'>"
+														.($i == $page ? "<span style='font-size: 1.1em; font-weight: 600;'>$i</span>" : $i)
+													."</a>"
+												."</li>";
+										}
+								$html.= "</ul>";
+							$html.= "</div>";
+						}
+						
+						$html.='<center><table style="border-color:#666666;" border=1 cellspacing=0 cellpadding=3>'
+							.'<thead>'
+							.'<tr>'
+								.'<th>'.$txt['cleantalk_check_users_tbl_select'].'</th>'
+								.'<th>'.$txt['cleantalk_check_users_tbl_username'].'</th>'
+								.'<th>'.$txt['cleantalk_check_users_tbl_joined'].'</th>'
+								.'<th>E-mail</th>'
+								.'<th>IP</th>'
+								.'<th>'.$txt['cleantalk_check_users_tbl_lastvisit'].'</th>'
+								.'<th>'.$txt['cleantalk_check_users_tbl_posts'].'</th>'
+							.'</tr>'
+							.'</thead>'
+							.'<tbody>';
+							
+						$found=false;
+						
+						while($row = $smcFunc['db_fetch_assoc'] ($result)){
 
-						$found=true;
-						$html.="<tr>
-						<td><input type='checkbox' name=ct_del_user[".$row['id_member']."] value='1' /></td>
-						<td>{$row['member_name']}&nbsp;<sup><a href='index.php?action=profile;u={$row['id_member']}' target='_blank'>{$txt['cleantalk_check_users_tbl_username_details']}</a></sup></td>
-						<td>".date("Y-m-d H:i:s",$row['date_registered'])."</td>
-						<td><a target='_blank' href='https://cleantalk.org/blacklists/".$row['email_address']."'><img src='https://cleantalk.org/images/icons/external_link.gif' border='0'/> ".$row['email_address']."</a></td>
-						<td><a target='_blank' href='https://cleantalk.org/blacklists/".$row['member_ip']."'><img src='https://cleantalk.org/images/icons/external_link.gif' border='0'/> ".$row['member_ip']."</a></td>
-						<td>".date("Y-m-d H:i:s",$row['last_login'])."</td>
-						<td style='text-align: center;'>{$row['posts']}&nbsp;<sup><a href='index.php?action=profile;area=showposts;u={$row['id_member']}' target='_blank'>{$txt['cleantalk_check_users_tbl_posts_show']}</a></sup></td>
-						</tr>";
+							$found=true;
+							$html.="<tr>
+							<td><input type='checkbox' name=ct_del_user[".$row['id_member']."] value='1' /></td>
+							<td>{$row['member_name']}&nbsp;<sup><a href='index.php?action=profile;u={$row['id_member']}' target='_blank'>{$txt['cleantalk_check_users_tbl_username_details']}</a></sup></td>
+							<td>".date("Y-m-d H:i:s",$row['date_registered'])."</td>
+							<td><a target='_blank' href='https://cleantalk.org/blacklists/".$row['email_address']."'><img src='https://cleantalk.org/images/icons/external_link.gif' border='0'/> ".$row['email_address']."</a></td>
+							<td><a target='_blank' href='https://cleantalk.org/blacklists/".$row['member_ip']."'><img src='https://cleantalk.org/images/icons/external_link.gif' border='0'/> ".$row['member_ip']."</a></td>
+							<td>".date("Y-m-d H:i:s",$row['last_login'])."</td>
+							<td style='text-align: center;'>{$row['posts']}&nbsp;<sup><a href='index.php?action=profile;area=showposts;u={$row['id_member']}' target='_blank'>{$txt['cleantalk_check_users_tbl_posts_show']}</a></sup></td>
+							</tr>";
+							
+						}
 						
+						$html.="</tbody></table></center>";
+						
+						if($pages && $pages != 1){
+							$html.= "<div style='margin: 10px;'>"
+									."<b>".$txt['cleantalk_check_users_pages'].":</b>"
+									."<ul style='display: inline-block; margin: 0; padding: 0;'>";
+										for($i = 1; $i <= $pages; $i++){
+											$html.= "<li style='display: inline-block;	margin-left: 10px;'>"
+													."<a href='".preg_replace('/(&ctcheckspam=.*|&spam_page=.*)/', '', $_SERVER['REQUEST_URI'])."&spam_page=$i'>"
+														.($i == $page ? "<span style='font-size: 1.1em; font-weight: 600;'>$i</span>" : $i)
+													."</a>"
+												."</li>";
+										}
+								$html.= "</ul>";
+							$html.= "</div>";
+						}
+						
+						$html.="<br /><center><input type=submit name='ct_delete_checked' value='".$txt['cleantalk_check_users_tbl_delselect']."' onclick='return confirm(\"".$txt['cleantalk_check_users_confirm']."\")'> <input type=submit name='ct_delete_all' value='".$txt['cleantalk_check_users_tbl_delall']."' onclick='return confirm(\"".$txt['cleantalk_check_users_confirm']."\")'><br />".$txt['cleantalk_check_users_tbl_delnotice']."<br /><br /></center>";
 					}
-					
-					$html.="</tbody></table></center>";
-					
-					if($pages && $pages != 1){
-						$html.= "<div style='margin: 10px;'>"
-								."<b>".$txt['cleantalk_check_users_pages'].":</b>"
-								."<ul style='display: inline-block; margin: 0; padding: 0;'>";
-									for($i = 1; $i <= $pages; $i++){
-										$html.= "<li style='display: inline-block;	margin-left: 10px;'>"
-												."<a href='".preg_replace('/(&ctcheckspam=.*|&spam_page=.*)/', '', $_SERVER['REQUEST_URI'])."&spam_page=$i'>"
-													.($i == $page ? "<span style='font-size: 1.1em; font-weight: 600;'>$i</span>" : $i)
-												."</a>"
-											."</li>";
-									}
-							$html.= "</ul>";
-						$html.= "</div>";
-					}
-					
-					$html.="<br /><center><input type=submit name='ct_delete_checked' value='".$txt['cleantalk_check_users_tbl_delselect']."' onclick='return confirm(\"".$txt['cleantalk_check_users_confirm']."\")'> <input type=submit name='ct_delete_all' value='".$txt['cleantalk_check_users_tbl_delall']."' onclick='return confirm(\"".$txt['cleantalk_check_users_confirm']."\")'><br />".$txt['cleantalk_check_users_tbl_delnotice']."<br /><br /></center>";
+				}else{
+					$html.='<center><div><b>'.$txt['cleantalk_check_users_key_is_bad'].'</b></div><br><br></center>';
 				}
 		    }
 		
@@ -1047,19 +1051,22 @@ function cleantalk_buffer($buffer)
 		$cleantalk_key_html = '';
 		
 		if(!isset($modSettings['cleantalk_api_key']))
-			$modSettings['cleantalk_api_key'] == '';
+			$modSettings['cleantalk_api_key'] = '';
 		
 		$cleantalk_key_html .= '<input type="text" name="cleantalk_api_key" id="cleantalk_api_key" value="'.$modSettings['cleantalk_api_key'].'" class="input_text">';
 		
 		if(!isset($modSettings['cleantalk_api_key_is_ok'])){
 			
-			$result = cleantalk\antispam\CleantalkHelper::noticeValidateKey($modSettings['cleantalk_api_key']);		
-			$result = ($result != false ? json_decode($result, true): null);
+			$result = CleantalkHelper::noticeValidateKey($modSettings['cleantalk_api_key']);		
 			
-			if($result && isset($result['valid']) && $result['valid'] == '1')
-				updateSettings(array('cleantalk_api_key_is_ok' => '1'), true);
-			else
-				updateSettings(array('cleantalk_api_key_is_ok' => '0'), true);
+			if(empty($result['error'])){
+				
+				if(isset($result['valid']) && $result['valid'] == '1')
+					updateSettings(array('cleantalk_api_key_is_ok' => '1'), true);
+				else
+					updateSettings(array('cleantalk_api_key_is_ok' => '0'), true);
+					
+			}
 		}
 		
 		if($modSettings['cleantalk_api_key_is_ok'] == '0' || $modSettings['cleantalk_api_key'] == ''){
@@ -1067,7 +1074,7 @@ function cleantalk_buffer($buffer)
 			$cleantalk_key_html .= "&nbsp<span style='color: red;'>".$txt['cleantalk_key_not_valid']."</span>";
 			$cleantalk_key_html .= "<br><br><a target='_blank' href='https://cleantalk.org/register?platform=smf&email=".urlencode($user_info['email'])."&website=".urlencode($_SERVER['SERVER_NAME'])."&product_name=antispam'>
 					<input type='button' value='".$txt['cleantalk_get_access_manually']."' />
-				</a> or ";
+				</a> {$txt['cleantalk_get_access_key_or']} ";
 			$cleantalk_key_html .= '<input name="spbc_get_apikey_auto" type="submit" class="spbc_manual_link" value="'.$txt['cleantalk_get_access_automatically'].'" onclick="location.href=location.href.replace(\'&finishcheck=1\',\'\').replace(\'&ctcheckspam=1\',\'\').replace(\'&ctgetautokey=1\',\'\')+\'&ctgetautokey=1\';return false;"/>';
 			$cleantalk_key_html .= "<br/><br/>";
 			$cleantalk_key_html .= "<div style='font-size: 10pt; color: #666 !important'>" . sprintf($txt['cleantalk_admin_email_will_be_used'], $user_info['email']) . "</div>";
@@ -1080,7 +1087,7 @@ function cleantalk_buffer($buffer)
 			
 		}
 		
-		$buffer = preg_replace('/<input type="text" name="cleantalk_api_key" id="cleantalk_api_key" value="[\da-zA-Z]{0,}" class="input_text" \/>/', $cleantalk_key_html, $buffer, 1);
+		$buffer = preg_replace('/<input type="text" name="cleantalk_api_key" id="cleantalk_api_key" value="[\da-zA-Z]{0,}" class="input_text"\s?\/?>/',$cleantalk_key_html, $buffer, 1);
 		
 	}
