+++ languages/english-utf8.xml
@@ -13,6 +13,7 @@ $txt['cleantalk_name'] = 'Antispam by CleanTalk';
 $txt['cleantalk_settings'] = 'Antispam by CleanTalk Mod settings';
 $txt['cleantalk_api_key'] = 'Access key';
 $txt['cleantalk_get_access_manually'] = 'Get access key manually';
+$txt['cleantalk_get_access_key_or'] = 'or';
 $txt['cleantalk_get_access_automatically'] = 'Get access key automatically';
 $txt['cleantalk_license_agreement'] = 'License agreement';
 $txt['cleantalk_key_valid'] = 'Access key is valid';
@@ -22,6 +23,7 @@ $txt['cleantalk_admin_email_will_be_used'] = 'Admin e-mail (%s) will be used for
 $txt['cleantalk_first_post_checking'] = 'Enable antispam test for first post on board for Newly registered members';
 $txt['cleantalk_api_key_description'] = 'Extension home page: <a target="_blank" href="https://cleantalk.org/">http://cleantalk.org</a>';
 $txt['cleantalk_logging'] = 'Enable logging';
+$txt['cleantalk_logging_postinput'] = 'All plugin\'s actions will be added to the forum\'s log. (<a target="_blank" href="%s/index.php?action=admin;area=logs;sa=errorlog;desc">Admin -> Error log</a>)';
 $txt['cleantalk_tell_others'] = 'Tell others about CleanTalk';
 $txt['cleantalk_tell_others_postinput'] = 'Checking this box places a small link in the footer that lets others know what anti-spam tool protects your site.';
 $txt['cleantalk_tell_others_footer_message'] = '<a href="https://cleantalk.org/smf-anti-spam-mod">SMF spam</a> blocked by CleanTalk';
@@ -30,14 +32,16 @@ $txt['cleantalk_sfw_postinput'] = 'This option allows to filter spam bots before
 $txt['cleantalk_check_users'] = '%CLEANTALK_CHECK_USERS%';
 $txt['cleantalk_check_users_button'] = 'Check users for spam';
 $txt['cleantalk_check_users_button_after'] = 'Anti-spam by CleanTalk will check all users against blacklists database and show you senders that have spam activity on other websites.';
+$txt['cleantalk_check_users_key_is_bad'] = 'Can not access cloud database. Access key is bad.';
 $txt['cleantalk_check_users_nofound'] = 'No spam users found';
 $txt['cleantalk_check_users_tbl_select'] = 'Select';
-$txt['cleantalk_check_users_tbl_username'] = 'Username';
+$txt['cleantalk_check_users_tbl_username_details'] = 'Details';
 $txt['cleantalk_check_users_tbl_joined'] = 'Registered';
 $txt['cleantalk_check_users_tbl_lastvisit'] = 'Last visit';
 $txt['cleantalk_check_users_tbl_posts'] = 'Posts';
+$txt['cleantalk_check_users_tbl_posts_show'] = 'Show';
 $txt['cleantalk_check_users_pages'] = 'Pages';
-$txt['cleantalk_check_users_done'] = 'Done. All users checked vie blacklists database, see result below.';
+$txt['cleantalk_check_users_done'] = 'Done. All users checked via blacklists database, see result below.';
 $txt['cleantalk_check_users_tbl_delselect'] = 'Delete selected';
 $txt['cleantalk_check_users_tbl_delall'] = 'Delete all';
 $txt['cleantalk_check_users_tbl_delnotice'] = 'All topics and posts of selected users will be deleted, too.';
@@ -46,7 +50,7 @@ $txt['cleantalk_sfw_die_notice_ip'] = 'SpamFireWall is activated for your IP ';
 $txt['cleantalk_sfw_die_make_sure_js_enabled'] = 'To continue working with web site, please make sure that you have enabled JavaScript.';
 $txt['cleantalk_sfw_die_click_to_pass'] = 'Please click bellow to pass protection,';
 $txt['cleantalk_sfw_die_you_will_be_redirected'] = 'Or you will be automatically redirected to the requested page after 3 seconds.';
-$txt['cleantalk_sfw_die_no_file'] = 'You IP looks like spammer\'s IP';
+$txt['cleantalk_sfw_die_no_file'] = 'Your IP looks like spammer\'s IP';
 $txt['cleantalk_banner_renew_1'] = 'Please, renew your anti-spam license for %s.';
 $txt['cleantalk_banner_renew_2'] = 'next year';
 $txt['cleantalk_banner_trial_1'] = 'Cleantalk\'s cloud service trial period ends, please upgrade to %s!';
