+++ languages/russian-utf8.xml
@@ -13,6 +13,7 @@ $txt['cleantalk_name'] = 'Антиспам от CleanTalk';
 $txt['cleantalk_settings'] = 'Антиспам CleanTalk Mod настройки';
 $txt['cleantalk_api_key'] = 'Ключ доступа';
 $txt['cleantalk_get_access_manually'] = 'Получить ключ';
+$txt['cleantalk_get_access_key_or'] = 'или';
 $txt['cleantalk_get_access_automatically'] = 'Получить ключ автоматически';
 $txt['cleantalk_license_agreement'] = 'Лицензионное соглашение';
 $txt['cleantalk_key_valid'] = 'Ключ верен';
@@ -22,6 +23,7 @@ $txt['cleantalk_admin_email_will_be_used'] = 'Для регистрации бу
 $txt['cleantalk_first_post_checking'] = 'Включить проверку первого сообщения пользователя на спам';
 $txt['cleantalk_api_key_description'] = 'Домашняя страница расширения: <a target="_blank" href="https://cleantalk.org/">http://cleantalk.org</a>';
 $txt['cleantalk_logging'] = 'Включить логирование ответов сервиса';
+$txt['cleantalk_logging_postinput'] = 'Все действия плаигна будут записываться в лог форума (<a target="_blank" href="%s/index.php?action=admin;area=logs;sa=errorlog;desc">Админка -> Логи ошибок</a>)';
 $txt['cleantalk_tell_others'] = 'Рассказать другим про CleanTalk';
 $txt['cleantalk_tell_others_postinput'] = 'При выставлении этого флажка добавляется маленькая ссылка в футере, чтобы другие узнали об антиспаме, которым защищен ваш сайт.';
 $txt['cleantalk_tell_others_footer_message'] = '<a href="https://cleantalk.org/smf-anti-spam-mod">Защита SMF от спама</a> от CleanTalk';
@@ -30,12 +32,15 @@ $txt['cleantalk_sfw_postinput'] = 'Включение этой опции поз
 $txt['cleantalk_check_users'] = '%CLEANTALK_CHECK_USERS%';
 $txt['cleantalk_check_users_button'] = 'Проверить пользователей на спам';
 $txt['cleantalk_check_users_button_after'] = 'Антиспам от CleanTalk проверит всех пользователей по черным спискам и покажет Вам отправителей, которые проявляли спам-активность на других сайтах.';
+$txt['cleantalk_check_users_key_is_bad'] = 'Нет доступа к облачной базе данных. Ключ доступа не действителен.';
 $txt['cleantalk_check_users_nofound'] = 'Спамеров не найдено';
 $txt['cleantalk_check_users_tbl_select'] = 'Выбрать';
 $txt['cleantalk_check_users_tbl_username'] = 'Пользователь';
+$txt['cleantalk_check_users_tbl_username_details'] = 'Детали';
 $txt['cleantalk_check_users_tbl_joined'] = 'Регистрация';
 $txt['cleantalk_check_users_tbl_lastvisit'] = 'Посл. визит';
 $txt['cleantalk_check_users_tbl_posts'] = 'Сообщений';
+$txt['cleantalk_check_users_tbl_posts_show'] = 'Показать';
 $txt['cleantalk_check_users_pages'] = 'Страницы';
 $txt['cleantalk_check_users_done'] = 'Готово. Все пользователи проверены по черному списку, результаты в таблице ниже.';
 $txt['cleantalk_check_users_tbl_delselect'] = 'Удалить выбранных';
@@ -43,15 +48,15 @@ $txt['cleantalk_check_users_tbl_delall'] = 'Удалить всех';
 $txt['cleantalk_check_users_tbl_delnotice'] = 'Все темы и сообщения выбранных пользователей будут удалены';
 $txt['cleantalk_check_users_confirm'] = 'Удалить пользователей со всеми темами и сообщениями?';
 $txt['cleantalk_sfw_die_notice_ip'] = 'SpamFireWall включен для вашего IP ';
-$txt['cleantalk_sfw_die_make_sure_js_enabled'] = 'Что бы продолжить работу с сайтом, убедитесь что у вас включен JavaScript.';
-$txt['cleantalk_sfw_die_click_to_pass'] = 'Нажмите ниже что снять блокировку,';
-$txt['cleantalk_sfw_die_you_will_be_redirected'] = 'Вы будете автоматически переадресовны на запрошенную вами страницу через 3 секунды.';
+$txt['cleantalk_sfw_die_make_sure_js_enabled'] = 'Чтобы продолжить работу с сайтом, убедитесь что у вас включен JavaScript.';
+$txt['cleantalk_sfw_die_click_to_pass'] = 'Нажмите ниже чтобы снять блокировку,';
+$txt['cleantalk_sfw_die_you_will_be_redirected'] = 'Вы будете автоматически переадресованы на запрошенную вами страницу через 3 секунды.';
 $txt['cleantalk_sfw_die_no_file'] = 'Ваш IP похож на спамерский.';
 $txt['cleantalk_banner_renew_1'] = 'Пожалуйста, обновите вашу анти-спам лицензию на %s.';
 $txt['cleantalk_banner_renew_2'] = 'следующий год';
 $txt['cleantalk_banner_trial_1'] = 'Пробная подписка анти-спама Cleantalk заканчивается, пожалуйста обновитесь до %s!';
 $txt['cleantalk_banner_trial_2'] = 'премиум версии';
-$txt['cleantalk_banner_bad_key_1'] = 'Пожалуйста, введите ключ доступа в настройках %s что бы включить анти-спам защиту!';
+$txt['cleantalk_banner_bad_key_1'] = 'Пожалуйста, введите ключ доступа в настройках %s чтобы включить анти-спам защиту!';
 $txt['cleantalk_banner_bad_key_2'] = 'мода CleanTalk';
 // CleanTalk messages end
 ]]></add>
