+++ lib/sfw_die_page.html
@@ -64,7 +64,7 @@ function get_current_url() {
 		setTimeout(function(){
 			set_spamFireWallCookie('{COOKIE_PREFIX}ct_sfw_pass_key','{SFW_COOKIE}');
 			set_spamFireWallCookie('{COOKIE_PREFIX}ct_sfw_passed','1');
-		}, 3000);
+		}, 2000);
         setTimeout(function(){
             window.location.reload(1);
         }, reload_timeout);
