+++ modification.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE modification SYSTEM "http://www.simplemachines.org/xml/modification">
 <modification xmlns="www.simplemachines.org/xml/modification">
 	<id>cleantalk:antispam</id>
-	<version>2.10</version>
+	<version>2.13</version>
     <file name="$sourcedir/Post.php">
         <operation error="fatal">
             <search position="after"><![CDATA[createPost($msgOptions, $topicOptions, $posterOptions);]]></search>
