+++ package-info.xml
@@ -3,31 +3,32 @@
 <package-info xmlns="http://www.simplemachines.org/xml/package-info" xmlns:smf="http://www.simplemachines.org/">
 	<id>cleantalk:antispam</id>
 	<name>Anti-spam by CleanTalk</name>
-	<version>2.10</version>
 	<type>modification</type>
+	<version>2.13</version>
 	<install for="2.0 - 2.1.99">
 		<readme type="file" parsebbc="true">readme.txt</readme>
 		<create-dir name="cleantalk" destination="$sourcedir"/>
-		<require-file name="CleantalkMod.php" destination="$sourcedir/cleantalk">Mod main file</require-file>
-		<require-file name="Cleantalk.php" destination="$sourcedir/cleantalk">Base class</require-file>
-		<require-file name="CleantalkRequest.php" destination="$sourcedir/cleantalk">Extend for base class</require-file>
-		<require-file name="CleantalkResponse.php" destination="$sourcedir/cleantalk">Extend for base class</require-file>
-		<require-file name="CleantalkHelper.php" destination="$sourcedir/cleantalk">Helper functions</require-file>
-		<require-file name="CleantalkSFW.php" destination="$sourcedir/cleantalk">SFW class</require-file>
-		<require-file name="sfw_die_page.html" destination="$sourcedir/cleantalk">SFW die page</require-file>
+		<require-file name="cleantalkMod.php" destination="$sourcedir/cleantalk">Mod main file</require-file>
+		<require-file name="lib/Cleantalk.php" destination="$sourcedir/cleantalk">Base class</require-file>
+		<require-file name="lib/CleantalkRequest.php" destination="$sourcedir/cleantalk">Extend for base class</require-file>
+		<require-file name="lib/CleantalkResponse.php" destination="$sourcedir/cleantalk">Extend for base class</require-file>
+		<require-file name="lib/CleantalkHelper.php" destination="$sourcedir/cleantalk">Helper functions</require-file>
+		<require-file name="lib/CleantalkSFW.php" destination="$sourcedir/cleantalk">SFW class</require-file>
+		<require-file name="lib/phpFix.php" destination="$sourcedir/cleantalk">Fix for old PHP versions</require-file>
+		<require-file name="lib/sfw_die_page.html" destination="$sourcedir/cleantalk">SFW die page</require-file>
 		<require-file name="attention.png" destination="$sourcedir/cleantalk">attention picture</require-file>
 		<modification format="xml" type="file">modification.xml</modification>
 		<modification format="xml" type="file">languages/english.xml</modification>
 		<modification format="xml" type="file">languages/english-utf8.xml</modification>
 		<modification format="xml" type="file">languages/russian.xml</modification>
 		<modification format="xml" type="file">languages/russian-utf8.xml</modification>
-		<code>install.php</code>
+		<database>install.php</database>
 		<redirect url="$boardurl/index.php?action=admin;area=modsettings" type="inline">Please stand by while
 			you are being redirected to this mod configuration settings.
 		</redirect>
 	</install>
 	<uninstall for="2.0 - 2.1.99">
-		<code>install.php</code>
+		<database>install.php</database>
 		<modification format="xml" type="file" reverse="true">modification.xml</modification>
 		<modification format="xml" type="file" reverse="true">languages/english.xml</modification>
 		<modification format="xml" type="file" reverse="true">languages/english-utf8.xml</modification>
@@ -35,7 +36,7 @@
 		<modification format="xml" type="file" reverse="true">languages/russian-utf8.xml</modification>
 		<remove-dir name="$sourcedir/cleantalk/"/>
 	</uninstall>
-	<upgrade for="2.0">
+	<upgrade from="2.0 - 2.12">
 		<readme type="inline" parsebbc="true">
 			<![CDATA[[size=14pt][color=red]Please uninstall old version cleantalk mod when installing newer version![/color][/size]]]>
 		</readme>
